requirements.txt
Django>=4.2.0
python-decouple==3.8
reportlab==4.0.4
requests==2.31.0
Pillow==10.0.0
google-auth==2.22.0
phonenumbers==8.13.20
django-crispy-forms==2.0
whitenoise==6.5.0